refactor and update

+++ models/estimated_expiration_price_notification.go
@@ -1,7 +1,7 @@
 package models
 
 type EstimatedExpirationPriceNotification struct {
-	Seconds     int     `json:"seconds"`
+	Seconds     uint64  `json:"seconds"`
 	Price       float64 `json:"price"`
 	IsEstimated bool    `json:"is_estimated"`
 }
