+++ models/trade.go
@@ -1,9 +1,9 @@
 package models
 
 type Trade struct {
-	TradeSeq       int     `json:"trade_seq"`
+	TradeSeq       uint64  `json:"trade_seq"`
 	TradeID        string  `json:"trade_id"`
-	Timestamp      int64   `json:"timestamp"`
+	Timestamp      uint64  `json:"timestamp"`
 	TickDirection  int     `json:"tick_direction"`
 	Price          float64 `json:"price"`
 	Iv             float64 `json:"iv"`
